[Mod] Make installer generation scripts a bit more reliable.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@4777 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ mptrack/manual_generator/wiki.py
@@ -158,7 +158,8 @@ toc.write("""
 """)
 toc.close()
 
-subprocess.call(['htmlhelp/hhc.exe', '"html/OpenMPT Manual.hhp"'])
+if(subprocess.call(['htmlhelp/hhc.exe', '"html/OpenMPT Manual.hhp"']) != 1):
+    raise Exception("Something went wrong during manual creation!")
 try:
     os.remove('../../packageTemplate/html/OpenMPT Manual.chm')
 except OSError:
